+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release Notes for Shopify
 
+## Unreleased
+
+- It is now possible to throttle the API calls in an attempt to avoid Shopify API rate limiting.
+- Fixed a bug where syncing Shopify variants would be limited to 50. ([#115](https://github.com/craftcms/shopify/issues/115))
+- Added `craft\shopify\console\controllers\SyncController::$throttle`.
+- Added `craft\shopify\services\Products::$throttle`.
+- Added `craft\shopify\services\Products::$sleepSeconds`.
+
 ## 5.1.2 - 2024-04-24
 
 - Fixed a bug where syncing meta fields would cause Shopify API rate limiting.